Serve follower and collection-add notifications from the collection index instead of a new service
