+++ src/api/handlers/xrpc/notification/listCollectionAdds.ts
@@ -29,14 +29,14 @@ export const listCollectionAdds: XRPCMethod<QueryParams, void, OutputSchema> = {
   auth: true,
   handler: async ({ params, c }): Promise<XRPCResponse<OutputSchema>> => {
     const logger = c.get('logger');
-    const notificationService = c.get('services').notification;
+    const collectionService = c.get('services').collection;
     const user = c.get('user');
 
     if (!user?.did) {
       throw new AuthenticationError('Authentication required');
     }
 
-    if (!notificationService) {
+    if (!collectionService) {
       // Returning an empty page would be indistinguishable from nobody having
       // collected the user's work, so a switched-off feature would read as a
       // real answer.
@@ -52,7 +52,7 @@ export const listCollectionAdds: XRPCMethod<QueryParams, void, OutputSchema> = {
       cursor: params.cursor,
     });
 
-    const page = await notificationService.listCollectionAdds(user.did, {
+    const page = await collectionService.listCollectionAdds(user.did, {
       limit: params.limit,
       ...(params.cursor !== undefined ? { cursor: params.cursor } : {}),
     });
