+++ src/api/server.ts
@@ -46,7 +46,6 @@ import type { KnowledgeGraphService } from '../services/knowledge-graph/graph-se
 import type { LayersDataLinkService } from '../services/layers/data-link-service.js';
 import type { MetricsService } from '../services/metrics/metrics-service.js';
 import type { ContentReportService } from '../services/moderation/content-report-service.js';
-import type { NotificationService } from '../services/notification/notification-service.js';
 import type { IPDSRegistry } from '../services/pds-discovery/pds-registry.js';
 import type { PDSScanner } from '../services/pds-discovery/pds-scanner.js';
 import type { PDSSyncService } from '../services/pds-sync/sync-service.js';
@@ -304,7 +303,6 @@ export interface ServerConfig {
   /**
    * Notification service instance (optional).
    */
-  readonly notificationService?: NotificationService;
 
   /**
    * Collaboration service for indexing invites / acceptances and deriving
@@ -454,7 +452,6 @@ export function createServer(config: ServerConfig): Hono<ChiveEnv> {
       indexRetryWorker: config.indexRetryWorker,
       personalGraph: config.personalGraphService,
       collection: config.collectionService,
-      notification: config.notificationService,
       collaborationService: config.collaborationService,
       admin: config.adminService,
       backfillManager: config.backfillManager,
